+++ netbox_dhcp_kea_plugin/__init__.py
@@ -2,7 +2,7 @@
 
 __author__ = """Łukasz Polański"""
 __email__ = "wookasz@gmail.com"
-__version__ = "0.7.2"
+__version__ = "0.7.3"
 
 
 from netbox.plugins import PluginConfig